General Update
+ Missing Arrays
+ JavaSer Tokener
+ Implimentation

+++ src/api/de/mineorea/jos/array/JOSDoubleArray.java
@@ -1,6 +1,6 @@
 package de.mineorea.jos.array;
 
-public interface JOSDoubleArray extends JOSArray<Integer , JOSDoubleArray>{
+public interface JOSDoubleArray extends JOSArray<Double , JOSDoubleArray>{
 
 	public abstract double[] getPrimitiveValue();
 	public abstract double[] setPrimitiveValue(double[] value);
